+++ KudaHitamQuant_full-simple.py
@@ -264,7 +264,7 @@ class KudahitamCompressorHBBA:
     @torch.no_grad()
     def compress(self, states: torch.Tensor, offload: bool = True) -> dict:
         if isinstance(states, (list, tuple)): states = states[0]
-        dev = states.device; shape = states.shape; flat = states.reshape(-1, shape[-1])
+        dev = states.device; shape = states.shape; flat = states.reshape(-1, shape[-1]).half()
         cuda_ext = load_cuda_ext()
         if not self.is_calibrated:
             norm = torch.norm(flat, dim=-1, keepdim=True); rotated = fwht((flat.float() / (norm+1e-8)) * self.d.to(dev)); self._calibrate_hbba(rotated[:1024])
